extracted RuleFilterOptions as ParsableArguments for reuse in other commands
